feat(lotto): get earning rate of user by Lotto game and exit
구현 내용 : 사용자의 수익률을 출력하고 로또 게임을 종료하는 기능 구현

+++ src/main/java/lotto/model/Calculator.java
@@ -32,13 +32,13 @@ public class Calculator {
         return userRanks;
     }
 
-    public float calculateEarningRate() {
-        int earned = userRanks[1] * EARNING_OF_1ST_PRIZE
+    public double calculateEarningRate() {
+        double earned = userRanks[1] * EARNING_OF_1ST_PRIZE
                 + userRanks[2] * EARNING_OF_2ND_PRIZE
                 + userRanks[3] * EARNING_OF_3RD_PRIZE
                 + userRanks[4] * EARNING_OF_4TH_PRIZE
                 + userRanks[5] * EARNING_OF_5TH_PRIZE;
-        float earningRate = earned / userInvestment * 100;
+        double earningRate = earned / userInvestment * 100;
         return earningRate;
     }
 }
